// src/api/gen/_model/saveResultResponseMissingItem.gen.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * TMT API
 * 또맛또(TMT) API
 * OpenAPI spec version: v1
 */

export type SaveResultResponseMissingItem =
  (typeof SaveResultResponseMissingItem)[keyof typeof SaveResultResponseMissingItem];

export const SaveResultResponseMissingItem = {
  COMPANION_TAG: "COMPANION_TAG",
  POSITIVE_POINT_TAG: "POSITIVE_POINT_TAG",
  RATING: "RATING",
  CONTENT: "CONTENT",
} as const;